Add seperate listeners for phpunit 6/7
